% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{DET_maketree}
\alias{DET_maketree}
\title{Process DET output}
\usage{
DET_maketree(x)
}
\arguments{
\item{x}{a fitted tree (output of \code{\link[det11]{ml_det}})}
}
\description{
Process a fitted density estimation tree; return the modified data frame
and necessary components to plot the tree with igraph.
}
