% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ml_det}
\alias{ml_det}
\title{Density Estimation Trees (DET)}
\usage{
ml_det(
  dataset,
  verbose = TRUE,
  folds = 10L,
  volreg = FALSE,
  maxsize = 10L,
  minsize = 5L
)
}
\arguments{
\item{dataset}{(note: rows must be variables; columns are observations)}

\item{verbose}{print out tree diagnostics?}

\item{folds}{number of folds for k-fold cross validation}

\item{volreg}{use volume regularization?}

\item{maxsize}{vaximum number of points allowed in a leaf}

\item{minsize}{minimum number of points allowed in a leaf}
}
\value{
a list with two elements
\itemize{
  \item varimp: a vector of relative variable importances
  \item tree: a \code{data.frame} with the following values
  \itemize{
    \item ID: the unique ID of this node
    \item parent_ID: the ID of the parent (0 for the root node)
    \item parent_split_edge: what edge from the parent is this node a descendent of? (\code{NULL} for the root node)
    \item leaf: 0 if not a leaf, 1 if a leaf node
    \item variable: what variable was used to split at this node? (1-indexed; leaves have value -1 as there is no split below a leaf)
    \item ratio: ratio of data points contained at or below this node compared to the whole training set
    \item right: children nodes on the right split have values > than this value
    \item right: children nodes on the left split have values <= than this value
    \item density: the estimated density (exp(log(ratio) - logVolume))
    \item volume: the volume contained in this node
    \item size: the number of data points at or below this node
  }
}
}
\description{
In general the default values work fine for most applications. 
See \url{https://mlpack.org/papers/det.pdf} for guidance if you want to change them.
}
