% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ml_det}
\alias{ml_det}
\title{Density Estimation Trees (DET)}
\usage{
ml_det(dataset, verbose = TRUE, folds = 10L, volreg = FALSE,
  maxsize = 10L, minsize = 5L)
}
\arguments{
\item{dataset}{(note: rows must be variables; columns are observations)}

\item{verbose}{print out tree diagnostics?}

\item{folds}{number of folds for k-fold cross validation}

\item{volreg}{use volume regularization?}

\item{maxsize}{vaximum number of points allowed in a leaf}

\item{minsize}{minimum number of points allowed in a leaf}
}
\description{
In general the default values work fine for most applications. 
See \url{https://mlpack.org/papers/det.pdf} for guidance if you want to change them.
}
